remove reference to Configure: Kubernetes
This section effectively configures the pod to relabel its bind-mounted files, which include important files on the host which should not be relabeled unless the user knows exactly what will happen.

This is likely not the best advice for users running on SELinux-enabled Kubernetes clusters. We need to find an equivalent to OpenShift's SecurityContextConstraints which configure what SELinux label the container process must run as, instead of relabeling the file system. Until we have that guidance, I think we should remove this.